Remove unsafe rand based backend
This may have some justification for testing purposes, i.e. to get
determinism, but in code relying on cryptography this is unlikely to be
required, and I believe in that kind of circumstance it can be
reimplemented in the test without too much difficulty, both clarifying
the intent and avoiding any accidental use of this dangerous fallback in
production.